+++ priv/repo/migrations/20170415141210_add_unique_index_to_email_and_nickname.exs
@@ -2,7 +2,7 @@ defmodule Pleroma.Repo.Migrations.AddUniqueIndexToEmailAndNickname do
   use Ecto.Migration
 
   def change do
-    create_if_not_exists unique_index(:users, [:email])
-    create_if_not_exists unique_index(:users, [:nickname])
+    create_if_not_exists(unique_index(:users, [:email]))
+    create_if_not_exists(unique_index(:users, [:nickname]))
   end
 end