+++ priv/repo/migrations/20170416122418_add_avatar_object_to_users.exs
@@ -3,7 +3,7 @@ defmodule Pleroma.Repo.Migrations.AddAvatarObjectToUsers do
 
   def change do
     alter table(:users) do
-      add :avatar, :map
+      add(:avatar, :map)
     end
   end
 end