+++ priv/repo/migrations/20170506222027_add_unique_index_to_apid.exs
@@ -2,6 +2,6 @@ defmodule Pleroma.Repo.Migrations.AddUniqueIndexToAPID do
   use Ecto.Migration
 
   def change do
-    create_if_not_exists unique_index(:users, [:ap_id])
+    create_if_not_exists(unique_index(:users, [:ap_id]))
   end
 end