+++ priv/repo/migrations/20170719152213_add_follower_address_to_user.exs
@@ -3,13 +3,13 @@ defmodule Pleroma.Repo.Migrations.AddFollowerAddressToUser do
 
   def up do
     alter table(:users) do
-      add :follower_address, :string, unique: true
+      add(:follower_address, :string, unique: true)
     end
   end
 
   def down do
     alter table(:users) do
-      remove :follower_address
+      remove(:follower_address)
     end
   end
 end