+++ priv/repo/migrations/20170906152508_create_o_auth_token.exs
@@ -3,11 +3,11 @@ defmodule Pleroma.Repo.Migrations.CreateOAuthToken do
 
   def change do
     create_if_not_exists table(:oauth_tokens) do
-      add :app_id, references(:apps)
-      add :user_id, references(:users)
-      add :token, :string
-      add :refresh_token, :string
-      add :valid_until, :naive_datetime_usec
+      add(:app_id, references(:apps))
+      add(:user_id, references(:users))
+      add(:token, :string)
+      add(:refresh_token, :string)
+      add(:valid_until, :naive_datetime_usec)
 
       timestamps()
     end