+++ priv/repo/migrations/20170917120416_add_tag_index.exs
@@ -4,6 +4,12 @@ defmodule Pleroma.Repo.Migrations.AddTagIndex do
   @disable_ddl_transaction true
 
   def change do
-    create index(:activities, ["(data #> '{\"object\",\"tag\"}')"], concurrently: true, using: :gin, name: :activities_tags)
+    create(
+      index(:activities, ["(data #> '{\"object\",\"tag\"}')"],
+        concurrently: true,
+        using: :gin,
+        name: :activities_tags
+      )
+    )
   end
 end