+++ priv/repo/migrations/20180325172351_add_follower_address_index_to_users.exs
@@ -3,7 +3,7 @@ defmodule Pleroma.Repo.Migrations.AddFollowerAddressIndexToUsers do
 
   @disable_ddl_transaction true
   def change do
-    create index(:users, [:follower_address], concurrently: true)
-    create index(:users, [:following], concurrently: true, using: :gin)
+    create(index(:users, [:follower_address], concurrently: true))
+    create(index(:users, [:following], concurrently: true, using: :gin))
   end
 end