+++ priv/repo/migrations/20180429094642_create_lists.exs
@@ -3,13 +3,13 @@ defmodule Pleroma.Repo.Migrations.CreateLists do
 
   def change do
     create_if_not_exists table(:lists) do
-      add :user_id, references(:users, on_delete: :delete_all)
-      add :title, :string
-      add :following, {:array, :string}
+      add(:user_id, references(:users, on_delete: :delete_all))
+      add(:title, :string)
+      add(:following, {:array, :string})
 
       timestamps()
     end
 
-    create_if_not_exists index(:lists, [:user_id])
+    create_if_not_exists(index(:lists, [:user_id]))
   end
 end