+++ priv/repo/migrations/20190122153157_update_activity_visibility.exs
@@ -27,10 +27,8 @@ defmodule Pleroma.Repo.Migrations.UpdateActivityVisibility do
     """
 
     execute(definition)
-
   end
 
   def down do
-
   end
 end