+++ priv/repo/migrations/20190301101154_add_default_tags_to_user.exs
@@ -2,7 +2,7 @@ defmodule Pleroma.Repo.Migrations.AddDefaultTagsToUser do
   use Ecto.Migration
 
   def up do
-    execute "UPDATE users SET tags = array[]::varchar[] where tags IS NULL"
+    execute("UPDATE users SET tags = array[]::varchar[] where tags IS NULL")
   end
 
   def down, do: :noop